requirements/prod.txt
celery==5.0.0
Django==4.2.8
django-allauth==0.59.0
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-user-agents==0.4.0
djangorestframework==3.14.0
geoip2==4.8.0
openpyxl==3.1.2
pillow==10.1.0
python-decouple==3.8
segno==1.6.0
sorl-thumbnail==12.10.0
sqids==0.4.1
unidecode==1.3.7